Update config.fish to start tmux AFTER keychain is set up and logging in. This is important because tmux will inherit the environment as it exists at the time of launch. In my case - I was starting tmux *before* keychain was setup with ssh credentials. The error was manifesting such that the `/run/user/1000` directory wasn't set up. fzf-lua (and maybe other plugins in neovim) require that to be setup to store sockets and other important run time artifacts.
This commit runs keychain first - which prompts for a passphrase - so I
guess that creates enough time for the environment to initialize before
executing tmux.

I just need to keep this in mind in the future if neovim starts not
working due to weird issues like unable run `config` or something.