Update fish.config to wait for `/run/user/1000/` directory to be created.
`/run/user/1000/` is usually the user id for the first non-root user
created on most linux systems, so this works right now on my own
machine.

I should probably make this dynamic so it works for any user,
not just the first non-root user. This would make the config more
flexible if I have ever have to use this config on a system where my
user isn't 1000, but the works for right now.